+++ pyccel/ast/class_defs.py
@@ -6,7 +6,7 @@
 This module contains all types which define a python class which is automatically recognised by pyccel
 """
 
-from pyccel.ast.builtin_methods.set_methods import SetAdd, SetClear
+from pyccel.ast.builtin_methods.set_methods import SetAdd, SetClear, SetCopy
 from pyccel.ast.builtin_methods.list_methods import ListAppend, ListInsert, ListPop, ListClear
 
 
@@ -153,10 +153,9 @@ ListClass = ClassDef('list', class_type = NativeHomogeneousList(),
 
 SetClass = ClassDef('set', class_type=NativeHomogeneousSet(),
         methods=[
-            PyccelFunctionDef('add', func_class = SetAdd,
-                decorators = {}),
-            PyccelFunctionDef('clear', func_class = SetClear,
-                decorators={}),
+            PyccelFunctionDef('add', func_class = SetAdd ),
+            PyccelFunctionDef('clear', func_class = SetClear),
+            PyccelFunctionDef('copy', func_class = SetCopy),
         ])
 
 #=======================================================================================